\name{MeTPeak-package}
\alias{MeTPeak-package}
\alias{MeTPeak}
\docType{package}
\title{
MeTPeak
}
\description{
The package is developed for the analysis of affinity-based epitranscriptome shortgun sequencing data from MeRIP-seq (maA-seq). Please feel free to contact Dr. Xiaodong Cui <xdchoi@gmail.com> if you have any questions.
}
\details{
\tabular{ll}{
Package: \tab MeTPeak\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2016-01-10\cr
License: \tab GPL-2 \cr
}

}

\references{

Xiaodong Cui, Meng, Jia, Shaowu Zhang, Yidong Chen, and Yufei Huang. "A novel algorithm for calling mRNA m6A peaks by modeling biological replicates in MeRIP-seq data"

}

\keyword{ Statistical Inference }

\examples{
# For usage, please check the main function with: 
?metpeak
}
